Add a sample custom app
Add offset for San Francisco

Rename SBP port from geoid to sample

Fix mistakes in offset flag

Sample app updates

+ Update post_image.sh to log the location of the build images, also
update the script so that it's less noisy, exporting DEBUG to the
environment will restore the prior logging behavior.

+ Move the S83sample_daemon start script into the package directory so
it can be conditionally installed if the sample daemon is enabled.

+ Change monit to include all .monitrc files in the /etc/monitrc.d
directory.

+ Add awscli to the make docker image so that `make firmware` can be run
inside the docker image (this will lessen the number of things that need
to be installed in VirtualBox for SDK users).

+ Attempt to fix the bug where the monit control port would not start
during boot if the eth0 interface was not up yet... this required
bumping to monit version 5.21.0 (which required forking buildroot)--
there may be a simpler workaround for this by using a unix socket.

+ Fix bug in docker base image build script

+ Fix bug in docker-wipe target

+ Remove sbp_router.yml changes, instead we'll rebroadcast POS and
HEARTBEAT messages over UDP.

Revert buildroot changes, use monit workaround

Remove bogus comment line

Update README.md

Update copyright

Update copyright

Share AWS creds with docker build container

Finish updating example to include UDP rebroadcast

+ Move most log messages to sbp_log so it's easier to debug without SSH
access.

+ Finish fixing the docker-cp target

+ Strip BR2_EXTERNAL_piksi_buildroot_PATH so that paths will show up as
relative, this will make more sense to SDK users (as opposed to a full
path that's relative to the docker container's file structure).

+ Add a Python example of how to receive and process the data

+ Make the UDP rebroadcast pack an entire SBP message into one datagram

+ Add a few settings to demo how they work

+ Upgrade sbp_log so that it can format messages

Remove unnecessary deps

Remove unnecessary include

Support semi-persistence of command line history

Conditionally compile the sample daemon

Misc clean-ups

+++ board/piksiv3/post_image.sh
@@ -5,6 +5,12 @@ if [ -z "$HW_CONFIG" ]; then
   exit 1
 fi
 
+if [ -n "$DEBUG" ]; then
+  LOG_DEBUG=/dev/stdout
+else
+  LOG_DEBUG=/dev/null
+fi
+
 CFG=piksiv3_$HW_CONFIG
 GIT_STRING=$(git -C $BR2_EXTERNAL_piksi_buildroot_PATH describe --tags        \
                  --dirty --always --match 'v[0-9]*\.[0-9]*\.[0-9]*')
@@ -21,6 +27,10 @@ UBOOT_VERSION=$(grep "$UBOOT_VERSION_REGEX" $UBOOT_MK_PATH | sed "s/${UBOOT_VERS
 
 UBOOT_BASE_DIR=`find $BUILD_DIR -maxdepth 1 -type d -name "uboot_custom-${UBOOT_VERSION}"`
 
+DEV_BIN_PATH=$OUTPUT_DIR/PiksiMulti-DEV-$FILE_GIT_STRING.bin
+FAILSAFE_BIN_PATH=$OUTPUT_DIR/PiksiMulti-FAILSAFE-$FILE_GIT_STRING.bin
+PROD_BIN_PATH=$OUTPUT_DIR/PiksiMulti-$FILE_GIT_STRING.bin
+
 if [[ -z "${UBOOT_BASE_DIR}" ]]; then
   echo "ERROR: Could not find uboot directory" >&2
   exit 1
@@ -30,46 +40,100 @@ UBOOT_PROD_DIR=$UBOOT_BASE_DIR/build/${CFG}_prod
 UBOOT_FAILSAFE_DIR=$UBOOT_BASE_DIR/build/${CFG}_failsafe
 UBOOT_DEV_DIR=$UBOOT_BASE_DIR/build/${CFG}_dev
 
+LOG_FILE=$(mktemp)
+trap 'rm -rf $LOG_FILE' EXIT
+
 generate_dev() {
+
+  echo -n "Generating DEV firmware image image... "
+
+  # Ensure log is empty
+  true >$LOG_FILE
+
   $UBOOT_DEV_DIR/tools/mkimage                                                \
   -A arm -T firmware -C none -O u-boot -n "FPGA bitstream"                    \
   -d $FIRMWARE_DIR/piksi_fpga.bit                                             \
-  $OUTPUT_DIR/piksi_fpga.img
+  $OUTPUT_DIR/piksi_fpga.img                                                  \
+    | tee -a $LOG_FILE &>$LOG_DEBUG
+
+  if [[ ${PIPESTATUS[0]} -ne 0 ]]; then
+    cat $LOG_FILE
+    exit 1
+  fi
+
+  # Ensure log is empty
+  true >$LOG_FILE
 
   $UBOOT_DEV_DIR/tools/image_table_util                                       \
   --append --print --print-images                                             \
-  --out "$OUTPUT_DIR/PiksiMulti-DEV-$FILE_GIT_STRING.bin"                     \
+  --out "$DEV_BIN_PATH"                                                       \
   --name "DEV $GIT_STRING"                                                    \
   --timestamp $(date +%s)                                                     \
   --hardware v3_$HW_CONFIG                                                    \
   --image $UBOOT_DEV_DIR/spl/u-boot-spl-dtb.img --image-type uboot-spl        \
   --image $UBOOT_DEV_DIR/u-boot.img --image-type uboot                        \
-  --image $OUTPUT_DIR/piksi_fpga.img --image-type fpga
+  --image $OUTPUT_DIR/piksi_fpga.img --image-type fpga                        \
+    | tee -a $LOG_FILE &>$LOG_DEBUG
+
+  if [[ ${PIPESTATUS[0]} -ne 0 ]]; then
+    cat $LOG_FILE
+    exit 1
+  fi
 
   rm $OUTPUT_DIR/piksi_fpga.img
+
+  echo "done."
 }
 
 generate_prod() {
+
+  echo -n "Generating PROD firmware image image... "
+
+  # Ensure log is empty
+  true >$LOG_FILE
+
   $UBOOT_PROD_DIR/tools/image_table_util                                      \
   --append --print --print-images                                             \
-  --out "$OUTPUT_DIR/PiksiMulti-$FILE_GIT_STRING.bin"                         \
+  --out "$PROD_BIN_PATH"                                                      \
   --name "$GIT_STRING"                                                        \
   --timestamp $(date +%s)                                                     \
   --hardware v3_$HW_CONFIG                                                    \
   --image $UBOOT_PROD_DIR/spl/u-boot-spl-dtb.img --image-type uboot-spl       \
   --image $UBOOT_PROD_DIR/u-boot.img --image-type uboot                       \
-  --image $BINARIES_DIR/uImage.$CFG --image-type linux
+  --image $BINARIES_DIR/uImage.$CFG --image-type linux                        \
+    | tee -a $LOG_FILE &>$LOG_DEBUG
+
+  if [[ ${PIPESTATUS[0]} -ne 0 ]]; then
+    cat $LOG_FILE
+    exit 1
+  fi
+
+  echo "done."
 }
 
 generate_failsafe() {
+
+  echo -n "Generating FAILSAFE firmware image image... "
+
+  # Ensure log is empty
+  true >$LOG_FILE
+
   $UBOOT_FAILSAFE_DIR/tools/image_table_util                                  \
   --append --print --print-images                                             \
-  --out "$OUTPUT_DIR/PiksiMulti-FAILSAFE-$FILE_GIT_STRING.bin"                \
+  --out "$FAILSAFE_BIN_PATH"                                                  \
   --name "FSF $GIT_STRING"                                                    \
   --timestamp $(date +%s)                                                     \
   --hardware v3_$HW_CONFIG                                                    \
   --image $UBOOT_FAILSAFE_DIR/spl/u-boot-spl-dtb.img --image-type uboot-spl   \
-  --image $UBOOT_FAILSAFE_DIR/u-boot.img --image-type uboot
+  --image $UBOOT_FAILSAFE_DIR/u-boot.img --image-type uboot                   \
+    | tee -a $LOG_FILE &>$LOG_DEBUG
+
+  if [[ ${PIPESTATUS[0]} -ne 0 ]]; then
+    cat $LOG_FILE
+    exit 1
+  fi
+
+  echo "done."
 }
 
 mkdir -p $OUTPUT_DIR
@@ -92,3 +156,18 @@ else
   echo "*** NO FIRMWARE FILES FOUND, NOT BUILDING PRODUCTION IMAGE ***"
 fi
 
+# Strip absolute path in case we're inside the docker container
+if [[ -n "$BR2_EXTERNAL_piksi_buildroot_PATH" ]]; then
+  PROD_BIN_PATH=${PROD_BIN_PATH#${BR2_EXTERNAL_piksi_buildroot_PATH}/}
+  DEV_BIN_PATH=${DEV_BIN_PATH#${BR2_EXTERNAL_piksi_buildroot_PATH}/}
+  FAILSAFE_BIN_PATH=${FAILSAFE_BIN_PATH#${BR2_EXTERNAL_piksi_buildroot_PATH}/}
+fi
+
+bold=$(tput rev)
+normal=$(tput sgr0)
+
+echo
+echo -e "${bold}>>> PROD firmware image located at:${normal}\n\t$PROD_BIN_PATH"
+echo -e "${bold}>>> DEV firmware image located at:${normal}\n\t$DEV_BIN_PATH"
+echo -e "${bold}>>> FAILSAFE firmware image located at:${normal}\n\t$FAILSAFE_BIN_PATH"
+echo
